Arc support
pass through FlaggedFilePath types with flags set.

This only adds support for arc on or off for every kit (default: off).

use with-arc: true to turn arc on in the KitSpec

All kits are exclusively on or off and will be explicitly flagged in the KitDeps.xcproject in the build sources section of the target